+++ src/routes/categories/+page.server.ts
@@ -3,12 +3,16 @@ import { categories } from '$lib/server/db/schema';
 import { eq } from 'drizzle-orm';
 import type { PageServerLoad } from './$types';
 import { getCategories } from '$lib/util';
+import { fail } from '@sveltejs/kit';
 
 export const actions = {
 	delete: async (event) => {
+		// Delete a category
 		const formData = await event.request.formData();
 
 		const id = formData.get('id') || '';
+
+		if (!id) return fail(400);
 
 		await db.delete(categories).where(eq(categories.id, id.toString()));
 	}
